Move main code body into _impl.py
