+++ trio_parallel/__init__.py
@@ -2,7 +2,7 @@
 
 from ._version import __version__
 
-from ._worker_processes import (
+from ._impl import (
     to_process_run_sync as run_sync,
     BrokenWorkerError,
     current_default_process_limiter,
